[rpm2extents] Perform digest computation within the validator
The validator calls `rpmcliVerifySignaturesFD` which under the hood
performs `Fread`. Digests are computed/updated for each `Fread`.

This diffs takes advantage of that by initializing the digest before
calling `rpmcliVerifySignaturesFD`. Once `rpmcliVerifySignaturesFD` as
returned and the file has been read, the digests are available.

This saves us from spawning a `digestor` process, as well as performing
an extra file read within it.