[reflink] fix support for hardlinks
a `suffix` made of ";${tid}" is now used until the changes are commited. As such, when
linking the file, we should link to the suffixed file.

The `suffix` is currently internal only to rpmPackageFilesInstall and there is
no obvious way to use data from struct `filedata_s` without exposing it to
a bunch of other places such as rpmplugin*.
We already keep track of the first file index. We used to use this to fetch the
file name, but due to suffix, the file is not named correctly, e.g it
has the name we will want at the end, once the transaction is commited,
not the temporary name.
Instead of re-implementing a local suffix that may change over time as
the implementation evolves, we can store the name of the file we should link to
in the hash instead of the index, which we were then using to fetch the
file name.
When hardlinking, we can then retrieve the name of the target file
instead of the index, like we previously did, and hardlink to that
without any issues.
Add a test to validate that reflink can handle hardlinking.

Originally, the test added would fail with:

```
error: reflink: Unable to hard link /foo/hello -> /foo/hello-bar;6207219c due to No such file or directory
error: Plugin reflink: hook fsm_file_install failed
error: unpacking of archive failed on file /foo/hello-bar;6207219c:
cpio: (error 0x2)
error: hlinktest-1.0-1.noarch: install failed
./rpm2extents.at:114: exit code was 1, expected 0
499. rpm2extents.at:112: 499. reflink hardlink package
(rpm2extents.at:112): FAILED (rpm2extents.at:114)
```

After this change, the test passes.