[reflink] remove requirement for executable stack flag
reflink was calling `bsearch` with a nested function comparator which
make GCC require the executable stack flag (see `man execstack`).
selinux prevents the use of this flag:
```
error: Failed to dlopen /usr/lib64/rpm-plugins/reflink.so
/usr/lib64/rpm-plugins/reflink.so: cannot enable executable stack as
shared object requires: Permission denied
```

To fix this, either rpm could be granted the correct selinux permission,
but this would open up execstack for more the whole rpm process.
Fundamentally, this is happening because there is no re-entrant version
of `bsearch`. We could probably use a global variable and be done with
it given that each rpm is processed sequencially, but that contract may
not hold true for ever.
Here we are copying stdlib's `bsearch` and making it re-entrant by
allowing to pass an void * data parameter where we can pass the key
size.

After applying this patch, when reflink.o is installed, it has the executable
flag cleared:
```
- /usr/lib64/rpm-plugins/reflink.so
```