RPM CoW: test header export failure
Reading and re-exporting the headers with RPM library functions
(rpmWriteSignature, headerWrite) produces a corrupted RPM package for
package azure-cli. Installation of the package fails with the following
error:
	signature region trailer: BAD, tag 2896420 type -738197504 offset 0 count 0

The issue can be reproduced with rpmtestheaderexport program which
reads the headers, re-exports them and just copy the rest of the RPM file.
Usage:
	/usr/local/bin/rpmtestheaderexport < azure-cli-2.36.0-1.el8.x86_64.rpm > azure-cli-2.36.0-1.el8.x86_64.export.rpm
	dnf -y install  azure-cli-2.36.0-1.el8.x86_64.export.rpm