RPM CoW: test header export failure
Reading and re-exporting the signature header with rpmWriteSignature
produces a corrupted RPM package for package azure-cli. Installation
of the package fails with the following error:
	signature region trailer: BAD, tag 2896420 type -738197504 offset 0 count 0

The issue can be reproduced with rpmtestheaderexport program. The program
reads the lead, the signature header and re-exports them. The rest of the
RPM file is just copied.
Usage:
	/usr/local/bin/rpmtestheaderexport < azure-cli-2.36.0-1.el8.x86_64.rpm > azure-cli-2.36.0-1.el8.x86_64.export.rpm
	dnf -y install  azure-cli-2.36.0-1.el8.x86_64.export.rpm